Struct caffe::Net

pub struct Net {
    // some fields omitted
}

A wrapper around a caffe::Net

Methods

impl Net

fn new(path: &Path, phase: Phase) -> Net

Construct a caffe::Net from a serialized caffe::NetParameter and a Phase.

fn wrap(net: Unique<Struct_CaffeNet>) -> Net

Wrap an existing caffe::Net over FFI.

fn copy_trained_layers_from(&mut self, path: &Path)

Set the weights of the network from the binary serialized caffe::NetParameter.

fn blob(&mut self, name: &str) -> Blob

Extract a pointer to the blob in the net.

fn forward_prefilled(&mut self)

Run a forward pass on the prefilled input blobs.